+++ admin/inc/basic.php
@@ -1305,14 +1305,37 @@ function isDebug(){
 }
 
 /**
+ * Check if GS Legacy version is an Alpha release
+ *
+ * @since 2025.2
+ * @uses get_site_version()
+ * @return boolean True if an alpha release
+ */
+function isAlpha() {
+	$version = strtok(get_site_version(false), '+');
+	$parts = explode('-', $version, 2);
+	if (count($parts) < 2) {
+		return false;
+	}
+	return strpos(strtolower($parts[1]), 'alpha') === 0;
+}
+
+/**
  * Check if GS Legacy version is a Beta release
  *
  * @since 3.3.0
  * @since 2025.1 Return boolean value
- * @return boolean true if beta release
+ * @since 2025.2 Refactored
+ * @uses get_site_version()
+ * @return boolean True if a beta release
  */
-function isBeta(){
-	return strpos(get_site_version(false), 'b') !== false;
+function isBeta() {
+	$version = strtok(get_site_version(false), '+');
+	$parts = explode('-', $version, 2);
+	if (count($parts) < 2) {
+		return false;
+	}
+	return strpos(strtolower($parts[1]), 'beta') === 0;
 }
 
 /**